+++ CHANGELOG.md
@@ -4,6 +4,9 @@ All notable changes to the "vscode-hugo-runner" extension will be documented in 
 
 Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
 
-## [Unreleased]
+## v0.1.0
 
-- Initial release
+- Command: Install Hugo (Windows only currently)
+- Command: Check Version
+- Command: Run Hugo Server
+- Command: Stop Hugo Server